combn: make Parser heap-allocable; aware of how to free itself
After this change, a stack-allocated `Parser(Value)` can be shifted onto
heap storage. In this case, the `Parser(Value)` interface remains intact,
and the parent value (e.g. the `Literal` struct, which itself exposes a
`Parser(Value)` interface) is retained on the heap.

This enables a parser to produce a new parser at runtime, have it be
heap-allocated, and returned as a parser result itself.

The only downside to this approach is that deinitialization currently
means a single _heap allocated_ parser can only be given as input to one
other parser, or else a double free would occur. This can be solved in the
future if needed.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>