combn: engine: Results can be cloned; a ResultStream subscriptions can clone them via a hook
Parser result streams are memoized. This works fine for stack-allocated values, but if
a result needs to be heap allocated it produces a problem (who owns the memory?) caught by
the Zig allocator as a leak or double free: multiple parsers are interacting with the
values and each expects to free the values in the stream.

Because of this, we introduce a hook here which allows results to clone themselves (i.e.
on the heap or using reference counting) when it is desirable to do so, i.e. when the results
one gets back from invoking a parser are cached due to memoization.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>