combn: per-Result ownership
A `ResultStream` is now said to own all values in it, and upon deinit it will
invoke the `deinit` method of all values in the set.

A `Result` now has an independent ownership value, which can be used to signal
a result was borrowed from somewhere else (e.g. another parser) and should not
have `deinit` invoked on it. A new `toUnowned()` method can be used to convert
an owned result into an unowned one.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>