all: reference count Parsers themselves
This makes the implementation functionally correct, but does make the usage quite
ugly (to be addressed later.) It completes the migration to ResultCount-owned
results.

The challenge is that parsers need to deinitialize their children parsers in the
graph/grammar, and in doing so multiple parsers may reference the same child. It
is thus necessary to deduplicate the deinitializations of them, implemented here
as ref counting.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>