all: add recursive option to parser deinit
Parsers must always know how to destroy the results they produce. This
is by nature of e.g. `Reentrant` parsers needing retry logic.

If a parser itself produces parsers (such as is the case with the Zorex DSL,
parsing a DSL and producing/compiling at runtime the parsers needed to parse
the language described by that DSL), then _the results of the DSL parsers need
to be freeable by the DSL parsers_. If results are combined across parsers to
compose complex ones, e.g.:

* DSL:ParseGrammar -> GrammarParser
    * DSL:ParseNumber -> NumberParser
       * DSL:ParseInt -> IntParser

Then when `ParseGrammar` tries to free its results, it would free `GrammarParser`,
which would recursively free `NumberParser` and `IntParser`. Then when `NumberParser`
is freed, we'd encounter a double free as it tries to free itself and `IntParser`
(its child).

And so, in the context of _a parser producing parsers_, the parser needs to be able
to free the results it produced (the parsers `GrammarParser`, `NumberParser`, etc.)
without recursively freeing children parsers.

This adds a simple `recursive` flag to `Parser.deinit` which does exactly this.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>